module-1/account-bootstrap-service-catalog-repo/pyproject.toml
[tool.poetry]
name = "account-bootstrap-service-catalog-repo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.115.0"
aws-cdk-aws-lambda-python-alpha = "^2.115.0a0"
constructs = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
